Added deal and findset given two cards
deck.js added a this.board array for the deck that stores the current
12 cards in the deck. Implemented
a deal function that takes the top 12 cards
from the deck and puts them on the board.
added updateBoard which will add 3 cards to board when set is made.

set.js added findSetCardID function that takes two cards and returns the
id of the third card that would make a set with the given cards.