Implemented logic for selecting a set
set.js - added logic for onclick function for images so it will add them
to a selected array and then check when that array hits three elements
for the possibility of a set.
If there is a set it removes those cards from the board and updates with
three new cards from the deck.

deck.js - added a function to remove a card from the board array by
filtering it out.